refactor(pages): adding Post pages
