+++ gatsby-config.ts
@@ -68,7 +68,7 @@ export default {
                 allMarkdownRemark(
                   limit: 1000,
                   sort: { order: DESC, fields: [frontmatter___date] },
-                  filter: { frontmatter: { template: { eq: "post" }, draft: { ne: true } } }
+                  filter: { frontmatter: { template: { in: ["post"] }, draft: { ne: true } } }
                 ) {
                   edges {
                     node {
