+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { default as useSiteMetadata } from "./use-site-metadata";
-export { default as usePostCategoriesList } from "./use-categories-list";
-export { default as usePostTagsList } from "./use-tags-list";
+export { default as usePostCategoriesList } from "./Post/use-post-categories-list";
+export { default as usePostTagsList } from "./Post/use-post-tags-list";
 export { default as useTheme, themeAtomKey } from "./use-theme";
