+++ src/templates/Post/PostsTemplate/PostsTemplate.tsx
@@ -18,7 +18,7 @@ interface Props {
   pageContext: PageContext;
 }
 
-const IndexTemplate: React.FC<Props> = ({ data, pageContext }: Props) => {
+const PostsTemplate: React.FC<Props> = ({ data, pageContext }: Props) => {
   const { pagination } = pageContext;
   const { hasNextPage, hasPrevPage, prevPagePath, nextPagePath } = pagination;
 
@@ -41,7 +41,7 @@ const IndexTemplate: React.FC<Props> = ({ data, pageContext }: Props) => {
 };
 
 export const query = graphql`
-  query IndexTemplate($limit: Int!, $offset: Int!) {
+  query PostsTemplate($limit: Int!, $offset: Int!) {
     allMarkdownRemark(
       limit: $limit
       skip: $offset
@@ -77,4 +77,4 @@ export const Head: React.FC<Props> = ({ pageContext }) => {
   return <Meta title={pageTitle} description={subtitle} />;
 };
 
-export default IndexTemplate;
+export default PostsTemplate;
